Heavy refactoring of Workout.hs
Got rid of class constraint in data declaration, eliminating need for -XGADTs
Added difficulty field to Workout to have more than one arg to Workout contructor
> This will be put in the common info later, but for now I will leave it there
WorkoutType is now data instead of a class.
Operations on WorkoutType will now have to be have full coverage for all types, instead
of each type implementing methods of the class.
This looks much neater and I think it will be much easier to code with