started work on getting a difficulty from stdin
so far getDifficulty will read a line from stdin and return an IO Difficulty
corresponding to the entered text, agnostic of case or whitespace.
An association table is provided in Workout.hs mapping Strings to Workouts
to aid this.
However, I have not yet implemented a good action to take when the entered text
does not match a difficulty. Currently it defaults to Easy, but it will be changed
to loop until the user enters a valid Difficulty