Started parsing workout types from stdin
getWorkoutTypeFunc :: IO (IO WorkoutType) will read a line from
stdin, parse for the name of a WorkoutType (reading more lines if
the current one isn't valid) and return a function that will read
from stdin and return a WorkoutType

Again this is aided by an association list. The previous one was renamed as well.