Fixed duration representation, and ...
Duration for an exercise is now represented as an Int.
It is read in normal form as numbers separated as colons.
Distance workouts have second precision, sports have minute.
This can be abused, because checking the input string is basic
and only checks that each number separated by colons is < 60.

In addition, more support for entering ... to find out what you can
enter was added